% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anova.critpat}
\alias{anova.critpat}
\title{ANOVA table for criterion-related profile analysis}
\usage{
\method{anova}{critpat}(object, ...)
}
\arguments{
\item{object}{\code{critpat} object resulting from \code{cpa}}

\item{...}{additional arguments affecting the ANOVA table produced.}
}
\description{
Compute analysis of variance tables for the \code{critpat} object.
}
\seealso{
\code{\link{cpa}}
}

