% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacc2001}
\alias{bacc2001}
\title{Baccalaureate and Beyond Longitudinal Study 2000}
\format{A data frame with 1080 rows and 4 variables:
\describe{
  \item{stem}{Are you a STEM major? 1: yes; 0: no}
  \item{major}{College major}
  \item{gpa}{GPA}
  \item{satq}{SAT quantitative}
  \item{satv}{SAT verbal}
}}
\source{
\url{https://nces.ed.gov/pubsearch/pubsinfo.asp?pubid=2003174}
}
\usage{
bacc2001
}
\description{
Simulated data based on the Baccalaureate and Beyond Longitudinal Study 2000/2001 based on the values presented in Tables 1 and 2 in Davison & Davenport (unpublished).
}
\keyword{dataset}
