% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy_plot.R
\name{dstudy_plot}
\alias{dstudy_plot}
\title{Plots various D studies for one and two-facet design}
\usage{
dstudy_plot(x, unit, facets, g_coef = T, bw = F)
}
\arguments{
\item{x}{object of class \code{mer} returned from the \code{lme4} function}

\item{unit}{the name of the unit of measurement}

\item{facets}{A list of length 1 (one-facet) or 2 (two-facet). If two-facet plot, then the second facet is the grouping variable.}

\item{g_coef}{Plot the generalizability coefficent? If FALSE, plot the dependability coefficient.}

\item{bw}{logical. Make the plot black and white? Default is FALSE.}
}
\description{
Plots multiple d studies for one or two-facet design.
}
\examples{
## One facet
one_facet <- lmer(Score ~ (1 | Participants) + (1 | Items), data = efData)
dstudy_plot(one_facet, unit = "Participants", facets = list(Items = c(8, 20, 30, 40)))
dstudy_plot(one_facet, unit = "Participants", facets = list(Items = c(8, 20, 30, 40)), g_coef = F)

## Two facet
two_facet <- lmer(scores ~ (1 | students) + (1  | prompts) + (1 | raters) + (1 | students:prompts) +  (1 | students:raters) + (1 | prompts:raters), data = writing)
twofacet_gstudy <- gstudy(two_facet)
dstudy_plot(twofacet_gstudy, unit = "students", facets = list(raters = 1:4, prompts = 1:4))
dstudy_plot(twofacet_gstudy, unit = "students", facets = list(raters = 1:4, prompts = 1:4), g_coef = F)
}
\seealso{
\code{\link{dstudy}}
}
