\docType{data}
\name{EEGS}
\alias{EEGS}
\title{Entrance Examination for Graduate Studies}
\format{\describe{
   \item{Form1_Q1}{First form of Quantitative 1}
   \item{Form2_Q1}{Second form of Quantitative 1}
   \item{Form1_Q2}{First form of Quantitative 2}
   \item{Form2_Q2}{Second form of Quantitative 2}
   \item{Form1_V}{First form of Verbal}
   \item{Form2_V}{Second form of Verbal}
 }}
\description{
  The \code{EEGS} is a subset of the Entrance Examination
  for Graduate Studies. There are three subscores
  in EEGS: Quantitative 1, Quantitative 2, and Verbal. In
  order to show the utility of subscore reliability method
  in this package, each subtest was separated into two
  parallel forms.
}
\keyword{datasets}

