% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstudy.R
\name{gstudy}
\alias{gstudy}
\title{Generate the variance components from a G-study}
\usage{
gstudy(x, fixed = NULL)
}
\arguments{
\item{x}{An lmer fitted object}
}
\description{
Generates the variance components and the relative magnitude (percent of total variance) of each variance component.
}
\examples{
one_facet <- lmer(Score ~ (1 | Participants) + (1 | Items), data = efData)
two_facet <- lmer(scores ~ (1 | students) + (1  | prompts) + (1 | raters) + (1 | students:prompts) +  (1 | students:raters) + (1 | prompts:raters), data = writing)
gstudy(one_facet)
gstudy(two_facet)
# Two facet with prompts fixed
gstudy(two_facet, fixed = "prompts")
}
