% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interest}
\alias{interest}
\title{Fabricated cognitive, personality, and vocational interest inventory}
\format{A data frame with 250 rows and 33 variables:
\describe{
 \item{gender}{1 is female and 2 is male}
 \item{educ}{Years of education}
 \item{age}{Age, in years}
 \item{vocab}{Vocabulary test}
 \item{reading}{Reading comprehension}
 \item{sentcomp}{Sentence completion}
 \item{mathmtcs}{Mathematics}
 \item{geometry}{Geometry}
 \item{analyrea}{Analytical reasoning}
 \item{socdom}{Social dominance}
 \item{sociabty}{Sociability}
 \item{stress}{Stress reaction}
 \item{worry}{Worry scale}
 \item{impulsve}{Impulsivity}
 \item{thrillsk}{Thrill-seeking}
 \item{carpentr}{Carpentry}
 \item{forestr}{Forest ranger}
 \item{morticin}{Mortician}
 \item{policemn}{Police}
 \item{fireman}{Fireman}
 \item{salesrep}{Sales representative}
 \item{teacher}{Teacher}
 \item{busexec}{Business executive}
 \item{stockbrk}{Stock broker}
 \item{artist}{Artist}
 \item{socworkr}{Social worker}
 \item{truckdvr}{Truck driver}
 \item{doctor}{Doctor}
 \item{clergymn}{Clergyman}
 \item{lawyer}{Lawyer}
 \item{actor}{Actor}
 \item{archtct}{Architect}
 \item{landscpr}{Landscaper}
}}
\source{
\url{https://web.archive.org/web/20071102075513/http://psych.colorado.edu/~carey/Courses/PSYC7291/DataSets/Documentation/InterestDataDoc.txt}
}
\usage{
interest
}
\description{
The data come from a fabricated cognitive, personality, and vocational interested inventory. This data set can be used to demonstrate regression and structural equation modeling.
}
\keyword{dataset}
