\docType{data}
\name{IPMMc}
\alias{IPMMc}
\title{Inventory of Personality and Mood Manifestation}
\format{This data frame contains the following columns:
\describe{
 \item{A}{Anxiety}
 \item{H}{Hypochondriasis}
 \item{S}{Schizophrenia}
 \item{B}{Bipolar Disorder}
 \item{R}{The Neurotic versus Psychotic Criterion Variable, where Neurotic (= 1) or Psychotic (= 0)}
 }}
\source{
  Davison, M. L., & Davenport, E. C. (2002). Identifying
  criterion-related patterns of predictor scores using
  multiple regression. \emph{Psychological Methods, 7}(4),
  468-484.
}
\description{
  The \code{IPMMc} data frame has 6 rows and 5 columns. See
  Davison and Davenport (2002) for more information.
}
\references{
  Davison, M. L., & Davenport, E. C. (2002). Identifying
  criterion-related patterns of predictor scores using
  multiple regression. \emph{Psychological Methods, 7}(4),
  468-484.
}
\keyword{datasets}

