\docType{data}
\name{leisure}
\alias{leisure}
\title{Leisure Activity Rankings}
\source{
  Tabachnick, B. G., & Fidell, L. S. (1996). \emph{Using
  multivariate statistics} (3rd ed.). New York: Harper
  Collins.
}
\description{
  The \code{leisure} dataset includes leisure activity
  rankings for three different groups: politicians,
  administrators, and belly-dancers. Rankings are provided
  in three categories: Reading-TV, Dance-TV, and TV-Ski.
  See Tabachnik and Fidell (1996) for more details.
}
\examples{
\dontrun{
data(leisure)
}}
\keyword{datasets}

