\docType{data}
\name{mod_data}
\alias{mod_data}
\title{Moderated profile analysis dummy data}
\format{This data frame contains the following columns:
    \describe{
      \item{dv}{Dependent variable}
      \item{pred1}{Predictor variable 1}
      \item{pred2}{Predictor variable 2}
      \item{mod}{The moderator variable}
      }
    }
\description{
Randomly generated data to test the \code{mpa} function.
}
\source{
This data was randomly generated to demonstrate how to use the \code{mpa} function.
}
\seealso{
\code{\link{mpa}}
}
\keyword{datasets}