\docType{data}
\name{nutrient}
\alias{nutrient}
\title{USDA Women's Health Survey}
\format{\describe{
   \item{calcium}{Calcium amount}
   \item{iron}{Iron amount}
   \item{protein}{Protein amount}
   \item{a}{Vitamin A amount}
   \item{c}{Vitamin C amount}
 }}
\description{
  In 1985, the United States Department of Agriculture (USDA) commissioned a study of women's nutrition.
  Nutrient intake was measured for a random sample of 737 women aged 25-50 years. Five nutritional components were
  measured: calcium, iron, protein, vitamin A and vitamin C.
}
\keyword{datasets}

