% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.prof}
\alias{plot.prof}
\title{Plot profile}
\usage{
\method{plot}{prof}(x, ...)
}
\arguments{
\item{x}{\code{prof} object resulting from \code{pr}}

\item{...}{additional arguments affecting the plot produced.}
}
\description{
Plots the profile
}
\seealso{
\code{\link{pr}}
}

