% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.critpat}
\alias{print.critpat}
\title{Print the criterion-related profile analysis}
\usage{
\method{print}{critpat}(x, ...)
}
\arguments{
\item{x}{object of class \code{critpat}}

\item{...}{additional arguments affecting the output produced.}
}
\description{
Prints the generic output from a \code{critpat} object.
}

