% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.prof}
\alias{print.prof}
\title{Print the pattern and level reliability via profile analysis}
\usage{
\method{print}{prof}(x, ...)
}
\arguments{
\item{x}{object of class \code{prof}}

\item{...}{additional arguments affecting the output produced.}
}
\description{
Prints the generic output from a \code{prof} object.
}
\seealso{
\code{\link{pr}}
}

