% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.profg}
\alias{print.profg}
\title{Print a profile by group: testing parallel, coincidental, and level profiles}
\usage{
\method{print}{profg}(x, ...)
}
\arguments{
\item{x}{object of class \code{profg}}

\item{...}{additional arguments affecting the output produced.}
}
\description{
Prints the generic output from a \code{profg} object.
}
\seealso{
\code{\link{pbg}}
}

