\docType{package}
\name{profileR-package}
\alias{profileR-package}
\title{Profile Analysis of Multivariate Data in R}
\description{
  The package profileR provides a set of multivariate methods and data visualization 
  tools to implement profile analysis and cross-validation techniques described in 
  Davison & Davenport (2002) and Bulut (2013).
}
\author{
  Okan Bulut \email{okanbulut84@gmail.com}

  Christopher David Desjardins
  \email{cddesjardins@gmail.com}
}
\references{
  Bulut, O. (2013). \emph{Between-person and within-person
  subscore reliability: Comparison of unidimensional and
  multidimensional IRT models}. (Doctoral dissertation).
  University of Minnesota. University of Minnesota,
  Minneapolis, MN. (AAT 3589000).

  Davison, M. L., & Davenport, E. C. (2002). Identifying
  criterion-related patterns of predictor scores using
  multiple regression. \emph{Psychological Methods, 7}(4),
  468-484.

  Davison, M. L., Kim, S-K., & Close, C. W. (2009). Factor
  analytic modeling of within person variation in score
  profiles. \emph{Multivariate Behavioral Research, 44},
  668-87.
}
\keyword{package}

