\docType{data}
\name{PS}
\alias{PS}
\title{A Hypothetical Personality Scale from Davison, Kim, and Close (2009)}
\format{\describe{
		\item{Person}{Person ID}
		\item{NEU}{Neurotic scale score}
		\item{PSY}{Psychotic scale score}
		\item{CD}{Character disorder scale score}
	}}
\source{
  Davison, M. L., Kim, S-K., & Close, C. W. (2009). Factor
  analytic modeling of within person variation in score
  profiles. \emph{Multivariate Behavioral Research, 44},
  668-87.
}
\description{
  The \code{PS} shows score profiles of six respondents to
  a hypothetical personality scale. It includes three types
  of profile patterns: Linearly increasing, inverted V, and
  linearly decreasing.
}
\references{
  Davison, M. L., Kim, S-K., & Close, C. W. (2009). Factor
  analytic modeling of within person variation in score
  profiles. \emph{Multivariate Behavioral Research, 44},
  668-87.
}
\keyword{datasets}

