\docType{data}
\name{spouse}
\alias{spouse}
\title{Love and Marriage Survey for Spouses}
\format{\describe{
   \item{item1}{Question 1 with a score ranging from 1 to 5.}
   \item{item2}{Question 2 with a score ranging from 1 to 5.}
   \item{item3}{Question 3 with a score ranging from 1 to 5.}
   \item{item4}{Question 4 with a score ranging from 1 to 5.}
   \item{spouse}{Spouse type. It is either "Husband" or "Wife"}
 }}
\description{
  The \code{spouse} is a data set comes from a study of love and marriage. A sample of 30 husbands and their wives were asked to respond to the following questions: \itemize{
  \item Question 1: What is the level of passionate love you feel for your partner?
  \item Question 2: What is the level of passionate love that your partner feels for you?
  \item Question 3: What is the level of companionate love that you feel for your partner?
  \item Question 4: What is the level of companionate love that your partner feels for you?
}

The responses to all four questions are on a five-point Likert scale where 1 indicates "none at all" and 5 indicates "tremendous amount". 
}
\examples{
\dontrun{
data(spouse)
}}
\keyword{datasets}

