% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.critpat.R
\name{summary.critpat}
\alias{summary.critpat}
\title{Summary of criterion-related profile analysis}
\usage{
\method{summary}{critpat}(object, ...)
}
\arguments{
\item{object}{object of class \code{critpat}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Provides a summary of the criterion-related profile analysis
}
\seealso{
\code{\link{cpa}}
}

