% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.profg}
\alias{summary.profg}
\title{Summary of profile by group: testing parallel, coincidental, and level profiles}
\usage{
\method{summary}{profg}(object, ...)
}
\arguments{
\item{object}{object of class \code{profg}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Provides a summary of profile group
}
\seealso{
\code{\link{pbg}}
}

