% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anova.critpat.R
\name{anova.critpat}
\alias{anova.critpat}
\title{Anova Tables}
\usage{
\method{anova}{critpat}(object, ...)
}
\arguments{
\item{object}{an object containing the results returned by a model fitting \code{cpa}.}

\item{...}{additional objects of the same type.}
}
\description{
Computes an analysis of variance table for a criterion-related profile analysis
}
\seealso{
\code{\link{cpa}}
}

