% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.critpat}
\alias{plot.critpat}
\title{Plot criterion-related profile}
\usage{
\method{plot}{critpat}(x, ...)
}
\arguments{
\item{x}{\code{critpat} object resulting from \code{cpa}}

\item{...}{additional arguments affecting the plot produced.}
}
\description{
Plots the criterion-related level and pattern profiles for each observation
}
\seealso{
\code{\link{cpa}}
}

