% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.prof.R
\name{plot.prof}
\alias{plot.prof}
\title{Plots a pattern and level reliability}
\usage{
\method{plot}{prof}(x, ...)
}
\arguments{
\item{x}{an object returned from the \code{pr} function}

\item{...}{additional objects of the same type.}
}
\description{
Plots the pattern vs. level reliability returned from the \code{pr} function of class \code{prof}.
}
\seealso{
\code{\link{pr}}
}

