% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.critpat.R
\name{print.critpat}
\alias{print.critpat}
\title{Print a criterion-related profile analysis}
\usage{
\method{print}{critpat}(x, ...)
}
\arguments{
\item{x}{object of class \code{critpat} returned from the \code{cpa} function}

\item{...}{additional objects of the same type.}
}
\description{
Prints the default output from fitting the \code{cpa} function.
}
\seealso{
\code{\link{cpa}}
}

