% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.critpat}
\alias{summary.critpat}
\title{Summary of criterion-related profile analysis}
\usage{
\method{summary}{critpat}(object, ...)
}
\arguments{
\item{object}{object of class \code{critpat}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Provides a summary of the criterion-related profile analysis
}
\seealso{
\code{\link{cpa}}
}

